requirements.txt
streamlit>=1.35.0
pillow>=10.0.0
pyyaml
langchain
langgraph
langchain-core
langchain-openai
